Moved test config dir to test folder
- this config is only used for testing so it should live in the
  `test/` dir
